feat: need more money

See merge request dokerplp/booking!29

+++ src/main/java/com/example/booking/service/OrdersService.java
@@ -6,12 +6,16 @@ import com.example.booking.dto.OrderReadDto;
 import com.example.booking.dto.UserReadDto;
 import com.example.booking.entity.Order;
 import com.example.booking.entity.User;
+import com.example.booking.exception.NotEnoughRightsException;
 import com.example.booking.mapper.OrderEntityMapper;
 import com.example.booking.mapper.OrderReadMapper;
 import com.example.booking.repository.HotelRepository;
 import com.example.booking.repository.OrdersRepository;
 import com.example.booking.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +28,8 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class OrdersService {
 
+    private final WalletService walletService;
+    private final UserService userService;
     private final OrdersRepository ordersRepository;
     private final RoomService roomService;
     private final OrderReadMapper orderReadMapper;
@@ -41,11 +47,20 @@ public class OrdersService {
         if (roomService.isRoomFree(room, orderDto.getDateIn(), orderDto.getDateOut())) {
             Order order = orderEntityMapper.map(createOrderDto);
             ordersRepository.save(order);
+            walletService.bookRoom(order);
         }
     }
 
 
     public List<OrderReadDto> showOrders(long userId) {
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var user = userService.findUserByName(principal.getUsername())
+                .orElseThrow(NotEnoughRightsException::new);
+
+        if (!user.getId().equals(userId)) {
+            throw new NotEnoughRightsException();
+        }
+
         return ordersRepository.findOrdersByUser_Id(userId).stream()
                 .map(orderReadMapper::map)
                 .toList();
@@ -53,6 +68,17 @@ public class OrdersService {
 
     @Transactional
     public void cancelOrder(long orderId) {
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        var order = ordersRepository.findById(orderId)
+                        .orElseThrow(() -> new IllegalArgumentException("Order " + orderId + " not found"));
+        var user = userService.findUserByName(principal.getUsername())
+                .orElseThrow(NotEnoughRightsException::new);
+        if (!order.getUser().equals(user)) {
+            throw new NotEnoughRightsException();
+        }
+
+        walletService.cancelRoom(order);
         ordersRepository.deleteById(orderId);
     }
 }
